+++ tests/index.js
@@ -68,6 +68,35 @@ async function go(colors) {
   // =================
   // Ready for testing
 
+  // ============ logger() tests begin
+  // 'dog' doesn't log at silly
+  if (loggers.isLevelEnabled({ tags: 'silly', category: 'dog' })) throw new Error();
+
+  {
+    const count = unitTest.entries.length;
+    loggers.logger('dog');
+    loggers.logger('dog').silly('a');
+    if (count !== unitTest.entries.length) throw new Error();
+  }
+
+  // Now 'dog' logs at silly
+  loggers.logger('dog', loggers.child(['warn', 'goofy'], { dog: 'woof' }));
+  if (!loggers.isLevelEnabled({ tags: 'silly', category: 'dog' })) throw new Error();
+
+  // Check the category of a saved logger is the same as the category provided to Loggers.logger()
+  if (loggers.logger('dog').category() !== 'dog') throw new Error();
+
+  {
+    const count = unitTest.entries.length;
+    loggers.logger('dog').silly('a'); // logLevel is specified via silly()
+    if (count !== unitTest.entries.length) throw new Error();
+    loggers.logger('dog').log('a'); // Use the tags for dog's logger
+    if (count === unitTest.entries.length) throw new Error();
+    const entry = unitTest.entries[unitTest.console.entries.length - 1];
+    if (entry.data.dog !== 'woof') throw new Error();
+  }
+  // ============ logger() tests end
+
   // Specify the level
   {
     const count = unitTest.entries.length;
@@ -178,7 +207,7 @@ async function go(colors) {
   {
     logger.info(['extra'], null, null, 'dragon');
     const entry = unitTest.entries[unitTest.entries.length - 1];
-    if (!entry.category === 'dragon') throw new Error();
+    if (entry.category !== 'dragon') throw new Error();
     if (!entry.tags.includes('extra')) throw new Error();
   }
 
@@ -757,10 +786,10 @@ async function go(colors) {
 
   {
     // These values must be tweaked whenever more entries are logged
-    if (unitTest.entries.length !== 167 + 10 * hasCloudWatch) throw new Error(unitTest.entries.length);
+    if (unitTest.entries.length !== 169 + 10 * hasCloudWatch) throw new Error(unitTest.entries.length);
     const len = Object.keys(unitTest.groupIds).length;
     if (len !== 32) throw new Error(len);
-    if (unitTest.dataCount !== 106 + 10 * hasCloudWatch) throw new Error(unitTest.dataCount);
+    if (unitTest.dataCount !== 108 + 10 * hasCloudWatch) throw new Error(unitTest.dataCount);
   }
 
   if (!onRan) throw new Error();
@@ -786,10 +815,11 @@ async function test() {
   let error;
 
   try {
-    await go(false);
     await go(true);
+    await go(false);
+
     // eslint-disable-next-line no-console
-    console.log('Successful');
+    console.log('\x1b[32m\x1b[40m\u2713\x1b Pass\x1b[0m');
   } catch (err) {
     error = err;
   }